refactor: standardize usage of `u32` type for IDs and improve database type alignment
- Replace `i64`/`i32` casts with `u32` for ID fields across all SQL queries to align with schema updates.
- Update table definitions to use `INT UNSIGNED` for consistency with Rust data types.
- Adjust `FromRow` implementations to properly handle `u32` conversions.
- Add support for verifying server creation within the database post-insertion.

+++ src-actix/server/installed_mods/installed_mods_watcher.rs
@@ -51,7 +51,7 @@ impl ServerData {
                         let exists = if let Some(filename) = path.file_name() {
                             let count: i64 = sqlx::query_scalar(r#"select count(*) from installed_mods where filename = ? and server_id = ?"#)
                                 .bind(filename.to_string_lossy())
-                                .bind(self.id as i64)
+                                .bind(self.id as u32)
                                 .fetch_one(pool)
                                 .await?;
                             count > 0
@@ -90,7 +90,7 @@ impl ServerData {
                         let exists = if let Some(filename) = path.file_name() {
                             let count: i64 = sqlx::query_scalar(r#"select count(*) from installed_mods where filename = ? and server_id = ?"#)
                                 .bind(filename.to_string_lossy())
-                                .bind(self.id as i64)
+                                .bind(self.id as u32)
                                 .fetch_one(pool)
                                 .await?;
                             count > 0
